Adds basic flag implementation for checking notes
This change implements a minimal version of global flags that can be set by dialogue, mainly adding the flag ID into the dialogue representation and adding a global class that has a list of flags to consult.

Also fixes a bug with the dialogue system where it hid unnecessary buttons but didn't put them back, and adjusts the color scheme a bit.